+++ packages/component-header-footer/src/footer/index.styles.js
@@ -25,8 +25,8 @@ const StyledFooter = styled.footer`
   // Base Styles
   * {
     box-sizing: border-box;
-    font-family: Arial, Helvetica, "Nimbus Sans L", "Liberation Sans", FreeSans,
-      sans-serif;
+    font-family:
+      Arial, Helvetica, "Nimbus Sans L", "Liberation Sans", FreeSans, sans-serif;
     line-height: 1.5rem;
 
     a:focus,
@@ -202,14 +202,14 @@ const StyledFooter = styled.footer`
 
     .flex-footer {
       .accordion-item,
-      .accordion-header,
-      .accordion-body {
+      .footer-accordion-header,
+      .footer-accordion-body {
         background: var(--color-divider-darker);
         color: var(--color-base-white);
         border: 0;
       }
 
-      .accordion-header {
+      .footer-accordion-header {
         border-top: 1px solid var(--color-divider-lighter);
         padding-left: 0;
 
@@ -217,14 +217,14 @@ const StyledFooter = styled.footer`
         .h5 {
           margin: 0;
           a,
-          .accordion-button {
+          .footer-accordion-button {
             ${flexCenter}
             justify-content: space-between;
           }
         }
 
         a,
-        .accordion-button {
+        .footer-accordion-button {
           color: var(--color-base-gray-6);
           padding: 1.5rem 0;
           text-decoration: none;
@@ -235,6 +235,7 @@ const StyledFooter = styled.footer`
           cursor: pointer;
           text-align: inherit;
           width: 100%;
+          transition: background-color 0.15s ease;
 
           &:hover {
             background: var(--color-divider-darker);
@@ -243,11 +244,11 @@ const StyledFooter = styled.footer`
           svg,
           fa-chevron-up {
             margin-left: 1rem;
-            transition: 0.5s cubic-bezier(0.19, 1, 0.19, 1);
+            transition: 0.2s cubic-bezier(0.25, 0.46, 0.45, 0.94);
             margin-left: 0.5rem;
             font-size: 1rem;
 
-            &.open {
+            &.column-open {
               transform: rotate(180deg);
             }
           }
@@ -257,7 +258,7 @@ const StyledFooter = styled.footer`
           border-top: 0;
           padding: 0;
           a,
-          .accordion-button {
+          .footer-accordion-button {
             padding: 0;
             cursor: default;
           }
@@ -268,16 +269,16 @@ const StyledFooter = styled.footer`
         }
       }
 
-      .accordion-body {
+      .footer-accordion-body {
         display: none;
         overflow: hidden;
         padding: 0 0 0 1.5rem;
 
-        &.collapsing {
+        &.footer-collapsing {
           display: block;
           max-height: 0px;
         }
-        &.show {
+        &.footer-column-show {
           display: block;
           max-height: 1000px;
         }
@@ -311,7 +312,7 @@ const StyledFooter = styled.footer`
       .h5 {
         margin: 0 0 1.5rem 0;
       }
-      .accordion-header:has(.h5, h5) {
+      .footer-accordion-header:has(.h5, h5) {
         h5,
         .h5 {
           margin: 0 0 1.5rem 0;